+++ tests/mrs_Mrs_test.py
@@ -16,7 +16,7 @@ from delphin.mrs import Mrs
 #from delphin.mrs import simplemrs  # for convenience in later tests
 from delphin.exceptions import XmrsError
 
-sp = Pred.stringpred
+sp = Pred.surface
 
 
 # for convenience