+++ tests/mrs_xmrs_test.py
@@ -33,10 +33,10 @@ class TestXmrs():
         assert len(x.eps()) == 0
         # nodeid and pred
         with pytest.raises(XmrsError):
-            x.add_eps([(10000, Pred.stringpred('_v_v_rel'))])
+            x.add_eps([(10000, Pred.surface('_v_v_rel'))])
         assert len(x.eps()) == 0
         # nodeid, pred, and label (the minimum)
-        x.add_eps([(10000, Pred.stringpred('_v_v_rel'), 'h1')])
+        x.add_eps([(10000, Pred.surface('_v_v_rel'), 'h1')])
         # make sure it was entered correctly and is unchanged
         assert len(x.eps()) == 1
         assert x.eps()[0][0] == 10000
@@ -46,7 +46,7 @@ class TestXmrs():
 
         # nodeid, pred, label, and argdict
         x = Xmrs()
-        x.add_eps([(10000, Pred.stringpred('_v_v_rel'), 'h1', {})])
+        x.add_eps([(10000, Pred.surface('_v_v_rel'), 'h1', {})])
         assert len(x.eps()) == 1
         assert x.eps()[0][0] == 10000
         ep = x.ep(10000)
@@ -57,7 +57,7 @@ class TestXmrs():
 
         # cannot have more than one ep with the same nodeid
         with pytest.raises(XmrsError):
-            x.add_eps([(10000, Pred.stringpred('_n_n_rel'), 'h3', {})])
+            x.add_eps([(10000, Pred.surface('_n_n_rel'), 'h3', {})])
         assert len(x.eps()) == 1
 
     def test_add_hcons(self):
@@ -112,7 +112,7 @@ class TestXmrs():
         assert x.xarg == 'e0'
 
     def test_nodeid(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         x = Xmrs()
         with pytest.raises(KeyError):
             x.nodeid('e2')
@@ -128,7 +128,7 @@ class TestXmrs():
         assert x.nodeid('x4', quantifier=True) == 11
 
     def test_nodeids(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         x = Xmrs(eps=[(10, sp('_n_n_rel'), 'h3', {'ARG0': 'x4'})])
         assert x.nodeids() == [10]
         assert x.nodeids(ivs=['x4']) == [10]
@@ -143,7 +143,7 @@ class TestXmrs():
         assert sorted(x.nodeids(ivs=['x4'], quantifier=False)) == [10]
 
     def test_ep(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         x = Xmrs()
         with pytest.raises(TypeError):
             x.ep()
@@ -153,7 +153,7 @@ class TestXmrs():
         assert x.ep(10)[1] == sp('_n_n_rel')
 
     def test_eps(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         x = Xmrs()
         assert len(x.eps()) == 0
         x.add_eps([(10, sp('_n_n_rel'), 'h3', {'ARG0': 'x4'})])
@@ -236,7 +236,7 @@ class TestXmrs():
         assert len(x.icons(left='x7')) == 1
 
     def test_variables_and_properties(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         # variables can be passed in with properties
         x = Xmrs(vars={'x1':{'PERS':'3','NUM':'sg'}, 'e2':{'SF':'prop'}})
         assert len(x.variables()) == 2
@@ -287,11 +287,11 @@ class TestXmrs():
         # KeyError on bad nodeid
         with pytest.raises(KeyError): x.pred(10)
         # but otherwise preds can be retrieved by nodeid
-        x.add_eps([(10, Pred.stringpred('_n_n_rel'), 'h3', {'ARG0': 'x4'})])
+        x.add_eps([(10, Pred.surface('_n_n_rel'), 'h3', {'ARG0': 'x4'})])
         assert x.pred(10).string == '_n_n_rel'
 
     def test_preds(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         x = Xmrs(
             eps=[
                 (10, sp('_v_v_rel'), 'h3', {'ARG0': 'e2', 'ARG1': 'x4'}),
@@ -324,12 +324,12 @@ class TestXmrs():
 
     def test_label(self):
         # retrieve the label for a single ep, or KeyError if no such nodeid
-        x = Xmrs(eps=[(10, Pred.stringpred('_v_v_rel'), 'h3', {'ARG0': 'e2'})])
+        x = Xmrs(eps=[(10, Pred.surface('_v_v_rel'), 'h3', {'ARG0': 'e2'})])
         assert x.label(10) == 'h3'
         with pytest.raises(KeyError): x.label(11)
 
     def test_labels(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         # same as Xmrs.labels() but with a list of nodeids
         x = Xmrs(
             eps=[
@@ -347,7 +347,7 @@ class TestXmrs():
     def test_args(self):
         # return the argument dict of a nodeid, or KeyError for missing nodeid
         x = Xmrs(
-            eps=[(10, Pred.stringpred('_v_v_rel'), 'h3',
+            eps=[(10, Pred.surface('_v_v_rel'), 'h3',
                   {'ARG0': 'e2', 'ARG1': 'x4'})]
         )
         assert x.args(10) == {'ARG0': 'e2', 'ARG1': 'x4'}
@@ -356,11 +356,11 @@ class TestXmrs():
         x.args(10)['ARG1'] = 'x6'
         assert x.args(10)['ARG1'] == 'x4'
         # return empty arg dict for EP without specified args:
-        x = Xmrs(eps=[(10, Pred.stringpred('_v_v_rel'), 'h3')])
+        x = Xmrs(eps=[(10, Pred.surface('_v_v_rel'), 'h3')])
         assert x.args(10) == {}
 
     def test_outgoing_args(self):
-        sp = Pred.stringpred
+        sp = Pred.surface
         # Outgoing args are those that, from some start node, go to
         # another in some way. These ways include:
         # regular variable args
@@ -404,7 +404,7 @@ class TestXmrs():
     def test_incoming_args(self):
         # incoming_args() is like the reverse of outgoing_args(), but
         # now it's many-to-one instead of one-to-many
-        sp = Pred.stringpred
+        sp = Pred.surface
         x = Xmrs(
             eps=[
                 (10, sp('_v_v_rel'), 'h3', {'ARG0': 'e2'}),
@@ -479,7 +479,7 @@ class TestXmrs():
         assert 10001 not in x
         assert '10000' not in x
         assert '_v_v_rel' not in x
-        assert Pred.stringpred('_v_v_rel') not in x
+        assert Pred.surface('_v_v_rel') not in x
 
     def test_labelset(self): pass
     def test_labelset_heads(self): pass
@@ -493,9 +493,9 @@ class TestXmrs():
         with pytest.raises(XmrsError):
             Xmrs(hcons=[('h0', 'qeq', 'h1')]).is_connected()
         # just a pred is fine (even without ARG0)
-        x = Xmrs(eps=[(10, Pred.stringpred('_v_v_rel'), 'h1', {})])
+        x = Xmrs(eps=[(10, Pred.surface('_v_v_rel'), 'h1', {})])
         assert x.is_connected() == True
-        x = Xmrs(eps=[(10, Pred.stringpred('_v_v_rel'), 'h1',
+        x = Xmrs(eps=[(10, Pred.surface('_v_v_rel'), 'h1',
                        {'ARG0':'e2'})])
         assert x.is_connected() == True
         # disconnected top is fine
@@ -553,7 +553,7 @@ class TestXmrs():
         pass
 
     def test_subgraph(self):
-        nodes = [Node(1,Pred.stringpred('verb'))]
+        nodes = [Node(1,Pred.surface('verb'))]
         links = [Link(0,1,'','H')]
         graph = Dmrs(nodes,links)
         new_graph = graph.subgraph([1])